refactor: 조합된 TimeList 컴포넌트 분리
