+++ front/src/components/TimeList/index.tsx
@@ -1,20 +1,7 @@
-import { Fragment } from 'react';
-
-import Conditional from '@components/Conditional';
-import { getDateTime, getHourMinutes } from '@utils/date';
-import type { MultipleTime, TimeSchedule } from '@typings/domain';
+import { PropsWithRequiredChildren } from '@typings/utils';
 import * as S from './styles';
 
-interface TimeListProps {
-  children: React.ReactNode;
-}
-
-interface ScrollProps {
-  children: React.ReactNode;
-}
-
 interface TimeProps {
-  children: React.ReactNode;
   onClick: () => void;
   isPastTime?: boolean;
   isPossible?: boolean;
@@ -22,42 +9,25 @@ interface TimeProps {
   index?: number;
 }
 
-interface DividerProps {
-  children: React.ReactNode;
-}
-
-interface ControlsProps {
-  children: React.ReactNode;
-}
-
-interface SelectAllButtonProps {
-  children: React.ReactNode;
+interface ButtonProps {
   onClick: () => void;
 }
 
-interface ConfirmButtonProps {
-  children: React.ReactNode;
-  onClick: () => void;
-}
-
-interface SelectedTimeProps {
-  children: React.ReactNode;
-}
-
-interface ReservationButtonProps {
-  children: React.ReactNode;
-  onClick: () => void;
-}
-
-const TimeList = ({ children }: TimeListProps) => {
+const TimeList = ({ children }: PropsWithRequiredChildren) => {
   return <S.TimeListContainer>{children}</S.TimeListContainer>;
 };
 
-const Scroll = ({ children }: ScrollProps) => {
+const Scroll = ({ children }: PropsWithRequiredChildren) => {
   return <S.ScrollContainer>{children}</S.ScrollContainer>;
 };
 
-const Time = ({ children, isPossible, dateTime, index, ...props }: TimeProps) => {
+const Time = ({
+  children,
+  isPossible,
+  dateTime,
+  index,
+  ...props
+}: PropsWithRequiredChildren<TimeProps>) => {
   return (
     <S.TimeItem
       isPossible={isPossible}
@@ -70,27 +40,27 @@ const Time = ({ children, isPossible, dateTime, index, ...props }: TimeProps) =>
   );
 };
 
-const Controls = ({ children }: ControlsProps) => {
+const Controls = ({ children }: PropsWithRequiredChildren) => {
   return <S.Controls>{children}</S.Controls>;
 };
 
-const SelectAllButton = ({ children, ...props }: SelectAllButtonProps) => {
+const SelectAllButton = ({ children, ...props }: PropsWithRequiredChildren<ButtonProps>) => {
   return <S.SelectAllButton {...props}>{children}</S.SelectAllButton>;
 };
 
-const ConfirmButton = ({ children, ...props }: ConfirmButtonProps) => {
+const ConfirmButton = ({ children, ...props }: PropsWithRequiredChildren<ButtonProps>) => {
   return <S.ConfirmButton {...props}>{children}</S.ConfirmButton>;
 };
 
-const Divider = ({ children }: DividerProps) => {
+const Divider = ({ children }: PropsWithRequiredChildren) => {
   return <S.Divider>{children}</S.Divider>;
 };
 
-const SelectedTime = ({ children }: SelectedTimeProps) => {
+const SelectedTime = ({ children }: PropsWithRequiredChildren) => {
   return <div>{children}</div>;
 };
 
-const ReservationButton = ({ children, ...props }: ReservationButtonProps) => {
+const ReservationButton = ({ children, ...props }: PropsWithRequiredChildren<ButtonProps>) => {
   return (
     <button {...props} autoFocus>
       {children}
@@ -98,120 +68,12 @@ const ReservationButton = ({ children, ...props }: ReservationButtonProps) => {
   );
 };
 
-interface ScheduleTimeListProps {
-  data: TimeSchedule[];
-  onClickTime: (dateTime: string) => void;
-  onSelectAll: () => void;
-  onSubmit: () => Promise<void>;
-  isSelectedAll: boolean;
-}
-
-export const ScheduleTimeList = ({
-  data,
-  onClickTime,
-  onSelectAll,
-  onSubmit,
-  isSelectedAll,
-}: ScheduleTimeListProps) => {
-  return (
-    <TimeList>
-      <TimeList.Scroll>
-        {data?.map(({ id, dateTime, ...props }) => (
-          <TimeList.Time
-            key={id}
-            onClick={() => onClickTime(dateTime)}
-            isPastTime={new Date() > getDateTime(dateTime)}
-            {...props}
-          >
-            {getHourMinutes(dateTime)}
-          </TimeList.Time>
-        ))}
-      </TimeList.Scroll>
-      <TimeList.Controls>
-        <TimeList.SelectAllButton onClick={onSelectAll}>
-          {isSelectedAll ? '전체 해제' : '전체 선택'}
-        </TimeList.SelectAllButton>
-        <TimeList.ConfirmButton onClick={onSubmit}>확인</TimeList.ConfirmButton>
-      </TimeList.Controls>
-    </TimeList>
-  );
-};
-
-interface MultipleTimeListProps {
-  data: MultipleTime[];
-  onClickTime: (dateTime: string) => void;
-  onSubmit: () => Promise<void>;
-}
-
-export const MultipleTimeList = ({ data, onClickTime, onSubmit }: MultipleTimeListProps) => {
-  return (
-    <TimeList>
-      <TimeList.Scroll>
-        {data?.map(({ id, dateTime, ...props }) => (
-          <TimeList.Time key={id} onClick={() => onClickTime(dateTime)} {...props}>
-            {dateTime}
-          </TimeList.Time>
-        ))}
-      </TimeList.Scroll>
-      <TimeList.Controls>
-        <TimeList.ConfirmButton onClick={onSubmit}>확인</TimeList.ConfirmButton>
-      </TimeList.Controls>
-    </TimeList>
-  );
-};
-
-interface ReservationTimeListProps {
-  data: TimeSchedule[];
-  onClickTime: (id: number, isPossible?: boolean) => void;
-  selectedTimeId: number | null;
-  onClickReservation: (scheduleId: number) => void;
-}
-
-export const ReservationTimeList = ({
-  data,
-  onClickTime,
-  selectedTimeId,
-  onClickReservation,
-}: ReservationTimeListProps) => {
-  return (
-    <TimeList>
-      {data?.map(({ id, dateTime, isPossible, ...props }, index) => {
-        const time = getHourMinutes(dateTime);
-        return (
-          <Fragment key={id}>
-            <Conditional condition={selectedTimeId === id}>
-              <TimeList.Divider>
-                <TimeList.SelectedTime>{time}</TimeList.SelectedTime>
-                <TimeList.ReservationButton onClick={() => onClickReservation(id)}>
-                  예약하기
-                </TimeList.ReservationButton>
-              </TimeList.Divider>
-            </Conditional>
-
-            <Conditional condition={selectedTimeId !== id}>
-              <TimeList.Time
-                onClick={() => onClickTime(id, isPossible)}
-                isPossible={isPossible}
-                dateTime={time}
-                index={index}
-                {...props}
-              >
-                {time}
-              </TimeList.Time>
-            </Conditional>
-          </Fragment>
-        );
-      })}
-    </TimeList>
-  );
-};
-
 TimeList.Time = Time;
 TimeList.Scroll = Scroll;
 TimeList.Controls = Controls;
-TimeList.Divider = Divider;
 TimeList.SelectAllButton = SelectAllButton;
 TimeList.ConfirmButton = ConfirmButton;
+TimeList.Divider = Divider;
 TimeList.SelectedTime = SelectedTime;
 TimeList.ReservationButton = ReservationButton;
